requirements.txt
streamlit==1.41.1
plotly==5.6.0
folium==0.16.0
streamlit-folium==0.20.0
Pillow==9.0.1
inflection==0.5.1